renamed file 'new_template_builder' to 'workfile_template_builder'
